Moved UpperBar.java to other package (it was the only file left in its package)
