+++ Rekeningsysteem/src/main/java/org/rekeningsysteem/application/Root.java
@@ -9,7 +9,6 @@ import javafx.scene.layout.BorderPane;
 import javafx.stage.Screen;
 import javafx.stage.Stage;
 
-import org.rekeningsysteem.application.top.UpperBar;
 import org.rekeningsysteem.application.working.MainPane;
 import org.rekeningsysteem.properties.PropertiesWorker;
 import org.rekeningsysteem.properties.PropertyModelEnum;
